+++ src/app/components/views/antropometria/antro-delete/antro-delete.component.ts
@@ -1,7 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
+import { take } from 'rxjs';
 import { Antropometria } from 'src/app/models/antropometria';
 import { AntropometriaService } from 'src/app/services/antropometria.service';
+import { ConfirmDialogService } from 'src/app/services/confirm-dialog.service';
+import { MensagemService } from 'src/app/services/mensagem.service';
 
 @Component({
   selector: 'app-antro-delete',
@@ -31,7 +34,8 @@ export class AntroDeleteComponent implements OnInit {
   constructor(
     private service: AntropometriaService,
     private route: ActivatedRoute,
-    private router: Router
+    private router: Router,
+    private mensagemService: MensagemService,private confirmService: ConfirmDialogService
   ) { }
 
   ngOnInit(): void {
@@ -60,12 +64,24 @@ export class AntroDeleteComponent implements OnInit {
   }
 
     delete(): void {
-      this.service.delete(this.antropometria.id!).subscribe((resposta) => {
-        this.router.navigate(["antropometrias"]);
-        this.service.mensagem("Antropometria deletada com sucesso");
-      }, err => {
-        this.service.mensagem(err.error.error)
-      });
+      const result$ = this.confirmService.abrir("Deseja excluir esta antropometria?")
+      result$.asObservable()
+        .pipe(
+          take(1),
+        ).subscribe({
+          next: (value) => {
+            if (value) {
+              this.service.delete(this.antropometria.id!).subscribe((resposta) => {
+                this.router.navigate(["antropometrias"]);
+                this.mensagemService.add("Antropometria deletada com sucesso");
+              }, err => {
+                this.mensagemService.add(err.error.error)
+              });
+            }
+          },
+        })
+
+      
     }
 
     cancel(): void {
