+++ src/app/components/views/funcionario/funcionario-delete/funcionario-delete.component.ts
@@ -1,7 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
+import { take } from 'rxjs';
 import { Funcionario } from 'src/app/models/funcionario';
+import { ConfirmDialogService } from 'src/app/services/confirm-dialog.service';
 import { FuncionarioService } from 'src/app/services/funcionario.service';
+import { MensagemService } from 'src/app/services/mensagem.service';
 
 @Component({
   selector: 'app-funcionario-delete',
@@ -27,7 +30,8 @@ export class FuncionarioDeleteComponent implements OnInit {
   constructor(
     private service: FuncionarioService,
     private route: ActivatedRoute,
-    private router: Router
+    private router: Router,
+    private mensagemService: MensagemService,private confirmService: ConfirmDialogService
   ) { }
 
   ngOnInit(): void {
@@ -47,12 +51,23 @@ export class FuncionarioDeleteComponent implements OnInit {
   }
 
   delete(): void {
-    this.service.delete(this.funcionario.id!).subscribe((resposta) => {
-      this.router.navigate(['funcionarios'])
-      this.service.mensagem('Funcionario deletado com sucesso!')
-    }, err => {
-      this.service.mensagem(err.error.error)
-    })
+    const result$ = this.confirmService.abrir("Deseja excluir este funcionario?")
+    result$.asObservable()
+      .pipe(
+        take(1),
+      ).subscribe({
+        next: (value) => {
+          if (value) {
+            this.service.delete(this.funcionario.id!).subscribe((resposta) => {
+              this.router.navigate(['funcionarios'])
+              this.mensagemService.add('Funcionario deletado com sucesso!')
+            }, err => {
+              this.mensagemService.add(err.error.error)
+            })
+          }
+        },
+      })
+
   }
 
   cancel(): void {
